+++ pkg/temingo/sortPaths.go
@@ -10,13 +10,14 @@ import (
 
 // Takes the paths from FileList.Files and sorts them into one list per filetype
 // Order of returned lists: templatePaths, metaTemplatePaths, partialPaths, metaPaths, staticPaths
-func (engine *Engine) sortPaths(fileList fileIO.FileList) ([]string, []string, []string, []string, []string) {
+func (engine *Engine) sortPaths(fileList fileIO.FileList) ([]string, []string, []string, []string, []string, []string) {
 	var (
-		templatePaths     []string
-		metaTemplatePaths []string
-		partialPaths      []string
-		metaPaths         []string
-		staticPaths       []string
+		templatePaths        []string
+		metaTemplatePaths    []string
+		partialPaths         []string
+		metaPaths            []string
+		markdownContentPaths []string
+		staticPaths          []string
 	)
 
 	for _, filePath := range fileList.Files { // Check what type of file we have
@@ -40,6 +41,11 @@ func (engine *Engine) sortPaths(fileList fileIO.FileList) ([]string, []string, [
 			if engine.Verbose {
 				log.Println("Identified as meta file:", filePath)
 			}
+		} else if path.Base(filePath) == engine.MarkdownContentFilename { // Making it easier to filter through them later and exclude them from staticPaths - they are not static files that should be copied to the outputDir
+			markdownContentPaths = append(markdownContentPaths, filePath)
+			if engine.Verbose {
+				log.Println("Identified as markdown content file:", filePath)
+			}
 		} else {
 			staticPaths = append(staticPaths, filePath)
 			if engine.Verbose {
@@ -48,5 +54,5 @@ func (engine *Engine) sortPaths(fileList fileIO.FileList) ([]string, []string, [
 		}
 	}
 
-	return templatePaths, metaTemplatePaths, partialPaths, metaPaths, staticPaths
+	return templatePaths, metaTemplatePaths, partialPaths, metaPaths, markdownContentPaths, staticPaths
 }